requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.9.4,<4.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.9.3
django-redis>=5.0,<6.0
celery==5.2.7
gunicorn==20.1.0
django-cors-headers==3.13.0
drf-yasg>=1.20,<2.0
djangorestframework-simplejwt==5.3.1


